pyproject.toml
[project]
name = "ffl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "streamlit>=1.52.2",
]
